safety stream record and playback enabled
